+++ javalib/src/main/scala/java/lang/IEEE754Helpers.scala
@@ -1,6 +1,7 @@
 package java.lang
 
 import scalanative.unsafe._
+import scalanative.unsigned._
 import scalanative.libc.errno
 
 import scalanative.posix.errno.ERANGE
@@ -29,7 +30,7 @@ private[java] object IEEE754Helpers {
 
   private def bytesToCString(bytes: Array[scala.Byte], n: Int)(
       implicit z: Zone): CString = {
-    val cStr = z.alloc(n + 1) // z.alloc() does not clear bytes.
+    val cStr = z.alloc((n + 1).toUInt) // z.alloc() does not clear bytes.
 
     var c = 0
     while (c < n) {
