+++ javalib/src/main/scala/java/net/PlainSocketImpl.scala
@@ -372,7 +372,7 @@ private[net] class PlainSocketImpl extends SocketImpl {
       var sent = 0
       while (sent < count) {
         val ret = socket
-          .send(fd.fd, cArr + sent, count - sent, socket.MSG_NOSIGNAL)
+          .send(fd.fd, cArr + sent, (count - sent).toUInt, socket.MSG_NOSIGNAL)
           .toInt
         if (ret < 0) {
           throw new IOException("Could not send the packet to the client")
@@ -387,7 +387,7 @@ private[net] class PlainSocketImpl extends SocketImpl {
     if (shutInput) -1
 
     val bytesNum = socket
-      .recv(fd.fd, buffer.asInstanceOf[ByteArray].at(offset), count, 0)
+      .recv(fd.fd, buffer.asInstanceOf[ByteArray].at(offset), count.toUInt, 0)
       .toInt
     if (bytesNum <= 0) {
       if (errno.errno == EAGAIN || errno.errno == EWOULDBLOCK) {
